Fix nonzero base issues with MFS device code and sigquit.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@2020 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/dosext/mfs/mangle.h
@@ -53,7 +53,7 @@ typedef char pstring[1024];
 
 
 /* prototypes */
-extern unsigned long is_dos_device(const char *fname);
+extern unsigned int is_dos_device(const char *fname);
 extern void mangle_name_83(char *s, char *MangledMap);
 extern BOOL do_fwd_mangled_map(char *s, char *MangledMap);
 extern BOOL name_convert(char *Name,BOOL mangle);
